+++ kernel/include/mm/mm.h
@@ -21,7 +21,7 @@ void        kvminit();
 void        kvminithart();
 pagetable_t kvmmake();
 pte_t*      walk(pagetable_t pgtbl, uint64 va, int alloc);
-void        mappages(pagetable_t pgtbl, uint64 va, uint64 pa, uint64 sz, int flags);
+void        mappages(pagetable_t pgtbl, uint64 va, uint64 pa, uint64 sz, uint64 flags);
 uint64      walkaddr(pagetable_t pgtbl, uint64 va);
 pagetable_t uvmmake(uint64 trapframe);
 pagetable_t uvminit(uint64 trapframe, char* init_code, int sz);
